+++ CHANGELOG.md
@@ -35,6 +35,8 @@
 - Fix adding `__new__()` methods for all class `__del__()` methods to
   anticipate when a class instance is created but not instantiated before garbage collection.
 - Add documentation for dynamic loop autograd limitations
+- Conform to Python's syntax for function arguments when calling built-ins inside of kernels, thus extending support for keyword arguments
+- Implement the assignment operator for `wp.quat`
 
 ## [1.2.1] - 2024-06-14